gui/requirements.txt
kafka-python 
matplotlib
